requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
streamlit==1.35.0
streamlit-authenticator==0.3.2
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
plotly==5.22.0
pmdarima==2.0.4
scikit-learn==1.4.2
statsmodels==0.14.2
requests==2.32.3
pytest==8.2.2
scipy==1.13.1
numpy==1.26.4
schedule==1.2.2